Fix witness syntax for array value assignment (#456)
* Fix witness syntax for array value assignment

Btor2 witness format expects an array assignment to contains a pair of index and value

* Skip assigments for state vars that have been explicitly initialized at #0

* Split btor2 witness printer header

* Merge duplicate print_btor_vals functions

* Replace written indices vector with set to speed up common case

* Add missing time argument

---------
