+++ CMakeLists.txt
@@ -208,6 +208,7 @@ set(SOURCES
   "${PROJECT_SOURCE_DIR}/modifiers/prophecy_modifier.cpp"
   "${PROJECT_SOURCE_DIR}/modifiers/static_coi.cpp"
   "${PROJECT_SOURCE_DIR}/modifiers/op_abstractor.cpp"
+  "${PROJECT_SOURCE_DIR}/printers/btor2_witness_printer.cpp"
   "${PROJECT_SOURCE_DIR}/printers/vcd_witness_printer.cpp"
   "${PROJECT_SOURCE_DIR}/refiners/array_axiom_enumerator.cpp"
   "${PROJECT_SOURCE_DIR}/smt/available_solvers.cpp"
